dotCMS/core#20664 Apply design to suggestions (#1718)
* Reorder components and services

* Merge

* Add basic suggestion list

* Implement FocusableOption in the list item

* Add basic keyboard events with static list items

* Clean up

* Add keydown event

* Bring back focus to the editor after floating button click and add some types

* Add sb story for list item

* Keep the focus in the editor while moving with the keyboard in the suggestion
