+++ libs/block-editor/src/lib/demo/block-editor.component.stories.ts
@@ -8,15 +8,15 @@ import { OrderListModule } from 'primeng/orderlist';
 import { MenuModule } from 'primeng/menu';
 
 import { NgxTiptapModule } from '../ngx-tiptap.module';
-import { SuggestionsService } from '../services/suggestions.service';
+import { SuggestionsService } from '../extensions/services/suggestions.service';
 import { BlockEditorComponent } from './block-editor.component';
-import { SuggestionsComponent } from '../suggestions/suggestions.component';
+import { SuggestionsComponent } from '../extensions/components/suggestions/suggestions.component';
 
-import { ActionButtonComponent } from '../extentions/action-button/action-button.component';
-import { ContentletBlockComponent } from '../extentions/blocks/contentlet-block/contentlet-block.component';
+import { ActionButtonComponent } from '../extensions/components/action-button/action-button.component';
+import { ContentletBlockComponent } from '../extensions/blocks/contentlet-block/contentlet-block.component';
 
 export default {
-    title: 'BlockEditorComponent'
+    title: 'Block Editor'
 };
 
 export const primary = () => ({
