+++ libs/block-editor/src/lib/extensions/components/action-button/action-button.component.html
@@ -1,0 +1,1 @@
+<button>➕</button>
