+++ libs/block-editor/src/lib/extensions/components/action-button/action-button.component.scss
@@ -1,0 +1,3 @@
+:host {
+    display: block;
+}
