+++ libs/block-editor/src/lib/extensions/components/action-button/action-button.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'dotcms-action-button',
+    templateUrl: './action-button.component.html',
+    styleUrls: ['./action-button.component.scss']
+})
+export class ActionButtonComponent {
+}
