+++ libs/block-editor/src/lib/extensions/components/suggestion-list/suggestion-list.component.html
@@ -1,0 +1,1 @@
+<ng-content></ng-content>
