+++ libs/block-editor/src/lib/extensions/components/suggestions-list-item/suggestions-list-item.component.html
@@ -1,0 +1,1 @@
+<ng-content></ng-content>
