+++ libs/block-editor/src/lib/extensions/components/suggestions-list-item/suggestions-list-item.component.scss
@@ -1,0 +1,5 @@
+:host {
+    display: block;
+    border: solid 1px silver;
+    padding: 1rem;
+}
