+++ libs/block-editor/src/lib/extensions/components/suggestions/suggestions.component.html
@@ -1,0 +1,5 @@
+<dotcms-suggestion-list #list>
+    <dotcms-suggestions-list-item>Option 1</dotcms-suggestions-list-item>
+    <dotcms-suggestions-list-item>Option 2</dotcms-suggestions-list-item>
+    <dotcms-suggestions-list-item>Option 3</dotcms-suggestions-list-item>
+</dotcms-suggestion-list>
