+++ libs/block-editor/src/lib/extensions/components/suggestions/suggestions.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { SuggestionsComponent } from './suggestions.component';
+
+describe('SuggestionsComponent', () => {
+  let component: SuggestionsComponent;
+  let fixture: ComponentFixture<SuggestionsComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ SuggestionsComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(SuggestionsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
