+++ libs/block-editor/src/lib/extensions/services/suggestions.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SuggestionsService } from './suggestions.service';
+
+describe('SuggestionsService', () => {
+  let service: SuggestionsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SuggestionsService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
