--- libs/block-editor/src/lib/extentions/action-button/action-button.component.html
@@ -1,1 +1,0 @@
-<button>➕</button>
