--- libs/block-editor/src/lib/extentions/action-button/action-button.component.scss
@@ -1,3 +1,0 @@
-:host {
-    display: block;
-}
