--- libs/block-editor/src/lib/extentions/action-button/action-button.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'dotcms-action-button',
-    templateUrl: './action-button.component.html',
-    styleUrls: ['./action-button.component.scss']
-})
-export class ActionButtonComponent {
-}
