--- libs/block-editor/src/lib/extentions/blocks/contentlet-block/contentlet-block.component.html
@@ -1,10 +1,0 @@
-<!-- <img src="https://demo.dotcms.com/dA/${data.identifier}/100wh" /> -->
-<p-card>
-    <ng-template pTemplate="header">
-        Title: {{ data.title }}
-    </ng-template>
-    <p>I am a angular component that uses primeng</p>
-    <ng-template pTemplate="footer">
-        Inode: {{ data.inode }}
-    </ng-template>
-</p-card>
