+++ libs/block-editor/src/lib/ngx-tiptap.module.ts
@@ -4,16 +4,17 @@ import { EditorDirective } from './editor.directive';
 import { BubbleMenuDirective } from './bubble-menu.directive';
 import { DraggableDirective } from './draggable.directive';
 import { NodeViewContentDirective } from './node-view-content.directive';
-import { SuggestionsComponent } from './suggestions/suggestions.component';
 
 import { MenuModule } from 'primeng/menu';
 import { CommonModule } from '@angular/common';
 import { CardModule } from 'primeng/card';
-import { ContentletBlockComponent } from './extentions/blocks/contentlet-block/contentlet-block.component';
+import { ContentletBlockComponent } from './extensions/blocks/contentlet-block/contentlet-block.component';
 
-import { SuggestionsService } from './services/suggestions.service';
-import { SuggestionListComponent } from './suggestion-list/suggestion-list.component';
-import { ActionButtonComponent } from './extentions/action-button/action-button.component';
+import { SuggestionsService } from './extensions/services/suggestions.service';
+import { SuggestionsComponent } from './extensions/components/suggestions/suggestions.component';
+import { SuggestionListComponent } from './extensions/components/suggestion-list/suggestion-list.component';
+import { ActionButtonComponent } from './extensions/components/action-button/action-button.component';
+import { SuggestionsListItemComponent } from './extensions/components/suggestions-list-item/suggestions-list-item.component';
 
 @NgModule({
     imports: [CommonModule, CardModule, MenuModule],
@@ -24,6 +25,7 @@ import { ActionButtonComponent } from './extentions/action-button/action-button.
         NodeViewContentDirective,
         SuggestionsComponent,
         SuggestionListComponent,
+        SuggestionsListItemComponent,
         ContentletBlockComponent,
         ActionButtonComponent
     ],
