--- libs/block-editor/src/lib/suggestion-list/suggestion-list.component.html
@@ -1,1 +1,0 @@
-<p-menu [model]="items"></p-menu>
