--- libs/block-editor/src/lib/suggestion-list/suggestion-list.component.ts
@@ -1,10 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-    selector: 'dotcms-suggestion-list',
-    templateUrl: './suggestion-list.component.html',
-    styleUrls: ['./suggestion-list.component.scss']
-})
-export class SuggestionListComponent {
-    @Input() items = [];
-}
