--- libs/block-editor/src/lib/suggestions/suggestions.component.html
@@ -1,1 +1,0 @@
-<dotcms-suggestion-list [items]="items"></dotcms-suggestion-list>
